+++ tests/twg.test.ts
@@ -1,8 +1,4 @@
-/* eslint-disable @typescript-eslint/prefer-includes */
-/* eslint-disable @stylistic/quotes */
-/* eslint-disable @typescript-eslint/no-unnecessary-condition */
-/* eslint-disable no-constant-binary-expression */
-
+// oxlint-disable typescript/no-unnecessary-condition no-constant-binary-expression typescript/strict-boolean-expressions
 import { twg } from "src/index"
 
 describe("twg()", () => {
@@ -65,13 +61,18 @@ describe("twg()", () => {
     describe("Remote classes:", () => {
         const remoteSingleClass = ["class"]
         const remoteMultipleClasses = ["multiple classes"]
-        const remoteMultipleClassesWithVar = ["multiple classes", { var: "class" }]
+        const remoteMultipleClassesWithVar = [
+            "multiple classes",
+            { var: "class" }
+        ]
         const remoteMultipleClassesWithMultipleVars = [
             "multiple classes",
             { var: "class" },
             { var: "multiple classes" }
         ]
-        const remoteMultipleClassesWithObjectOnly = [{ var: "multiple classes" }]
+        const remoteMultipleClassesWithObjectOnly = [
+            { var: "multiple classes" }
+        ]
         const remoteMultipleClassesAndVarsWithObjectOnly = [
             {
                 var1: "multiple classes",
@@ -116,13 +117,16 @@ describe("twg()", () => {
 
         it.each([
             {
-                args: ["multiple classes", `bg-${isAndOr1 ? `white` : "black)"}`],
+                args: [
+                    "multiple classes",
+                    `bg-${isAndOr1 ? "white" : "black)"}`
+                ],
                 expected: "multiple classes bg-white"
             },
             {
                 args: [
                     "multiple classes",
-                    /abcd_\)/.test(cls) ? "bg-white" : "bg-black"
+                    /abcd_\)/u.test(cls) ? "bg-white" : "bg-black"
                 ],
                 expected: "multiple classes bg-black"
             },
@@ -180,7 +184,10 @@ describe("twg()", () => {
             },
             {
                 //*
-                args: ["multiple classes", { var: isAndOr1 && isAndOr2 && "class" }],
+                args: [
+                    "multiple classes",
+                    { var: isAndOr1 && isAndOr2 && "class" }
+                ],
                 expected: "multiple classes"
             },
             {
@@ -329,7 +336,10 @@ describe("twg()", () => {
             { args: [{ class: is_And_Or_2 && isAndOr1 }], expected: "" },
             { args: [{ class: is_And_Or_2 && isAndOr2 }], expected: "" },
             {
-                args: [{ class: isAndOr1 }, { "multiple classes": is_And_Or_2 }],
+                args: [
+                    { class: isAndOr1 },
+                    { "multiple classes": is_And_Or_2 }
+                ],
                 expected: "class"
             },
             {
@@ -439,7 +449,10 @@ describe("twg()", () => {
             {
                 args: [
                     {
-                        var1: ["multiple classes", { var2: "in array with var" }]
+                        var1: [
+                            "multiple classes",
+                            { var2: "in array with var" }
+                        ]
                     }
                 ],
                 expected:
@@ -457,7 +470,10 @@ describe("twg()", () => {
                 expected: "multiple classes var:class"
             },
             {
-                args: ["multiple classes", { var: "multiple classes with var" }],
+                args: [
+                    "multiple classes",
+                    { var: "multiple classes with var" }
+                ],
                 expected:
                     "multiple classes var:multiple var:classes var:with var:var"
             },
@@ -481,7 +497,10 @@ describe("twg()", () => {
                 expected: "var:class multiple classes"
             },
             {
-                args: [{ var: "multiple classes with var" }, "multiple classes"],
+                args: [
+                    { var: "multiple classes with var" },
+                    "multiple classes"
+                ],
                 expected:
                     "var:multiple var:classes var:with var:var multiple classes"
             },
@@ -670,7 +689,8 @@ describe("twg()", () => {
                     { var1: "class" },
                     { var2: "multiple classes" }
                 ],
-                expected: "multiple classes var1:class var2:multiple var2:classes"
+                expected:
+                    "multiple classes var1:class var2:multiple var2:classes"
             },
             {
                 args: [
@@ -731,7 +751,11 @@ describe("twg()", () => {
                 expected: "multiple classes var:class var:class"
             },
             {
-                args: ["multiple classes", { var: "class-1" }, { var: "class-2" }], //*
+                args: [
+                    "multiple classes",
+                    { var: "class-1" },
+                    { var: "class-2" }
+                ], //*
                 expected: "multiple classes var:class-1 var:class-2"
             },
             {
@@ -798,9 +822,7 @@ describe("twg()", () => {
         })
     })
 
-    /**
-     * @borrows https://github.com/Noriller/easy-tailwind/blob/master/src/index.spec.ts#L80C3-L120C6
-     */
+    /** @borrows https://github.com/Noriller/easy-tailwind/blob/master/src/index.spec.ts#L80C3-L120C6 */
     describe("Complex cases:", () => {
         /* cSpell:disable */
         it.each([
@@ -858,9 +880,7 @@ describe("twg()", () => {
         /* cSpell:enable */
     })
 
-    /**
-     * @borrows https://github.com/Noriller/easy-tailwind/blob/master/src/index.spec.ts#L122C3-L127C5
-     */
+    /** @borrows https://github.com/Noriller/easy-tailwind/blob/master/src/index.spec.ts#L122C3-L127C5 */
     describe("Falsy cases:", () => {
         const isUndefined = undefined
         const isNull = null